Split msg_handler.go into front_handlers and table_handlers
